\name{paramap-package}

\alias{paramap-package}

\docType{package}

\title{paramap}

\description{This package provides exploratory factor analysis-related functions for 
	assessing dimensionality. There are functions for seven different procedures 
	for determining the number of factors, including functions for parallel analysis 
	and the minimum average partial test. There are functions for conducting principal 
	components analysis, principal axis factor analysis, maximum likelihood factor 
	analysis, image factor analysis, and extension factor analysis, all of which can 
	take raw data or correlation matrices as input and with options for conducting 
	the analyses using Pearson correlations, Kendall correlations, Spearman correlations, 
	or polychoric correlations. Varimax rotation, promax rotation, and Procrustes 
	rotations can be performed. Additional functions focus on the factorability of 
	a correlation matrix, the congruences between factors from different datasets, and 
	for assessing local independence. 
}

