\name{paramap-package}

\alias{paramap-package}

\docType{package}

\title{ paramap }

\description{This package provides factor analysis-related functions for assessing 
	dimensionality. Users can request that the analyses be conducted using polychoric correlations, 
	which are preferable to Pearson correlations for item-level data, and a polychoric 
	correlation matrix is returned for possible further analyses. There are also functions 
	for conducting principal components analysis, principal axis 
	factor analysis, maximum likelihood factor analysis, image factor analysis, and extension 
	factor analysis, all of which can take raw data or correlation matrices as input and with 
	options for conducting the analyses using Pearson correlations, Kendall correlations, 
	Spearman correlations, or polychoric correlations. Varimax rotation, promax 
	rotation, and Procrustes rotations can be performed.
}

